requirement.txt
selenium==4.15.2
pytest==7.4.3
pytest-xdist==3.3.1
pytest-html==4.1.1
allure-pytest==2.13.2
openpyxl==3.1.2
pyyaml==6.0.1
webdriver-manager==4.0.1
colorlog==6.7.0
python-dotenv==1.0.0